Separating read models into folders/namespaces
