+++ Source/Admin/Read/HealthRiskFeatures/HealthRisk.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Read
+namespace Read.HealthRiskFeatures
 {
     public class HealthRisk
     {
