+++ Source/Admin/Read/HealthRiskFeatures/IHealthRisks.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Read
+namespace Read.HealthRiskFeatures
 {
     public interface IHealthRisks
     {
