+++ Source/Admin/Read/NationalSocietyFeatures/INationalSocieties.cs
@@ -1,11 +1,11 @@
-﻿/*---------------------------------------------------------------------------------------------
+/*---------------------------------------------------------------------------------------------
  *  Copyright (c) 2017 International Federation of Red Cross. All rights reserved.
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
 using System.Collections.Generic;
 
-namespace Read
+namespace Read.NationalSocietyFeatures
 {
     public interface INationalSocieties
     {
