+++ Source/Admin/Read/NationalSocietyFeatures/NationalSocieties.cs
@@ -1,4 +1,4 @@
-﻿/*---------------------------------------------------------------------------------------------
+/*---------------------------------------------------------------------------------------------
  *  Copyright (c) 2017 International Federation of Red Cross. All rights reserved.
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
@@ -6,7 +6,7 @@
 using MongoDB.Driver;
 using System.Collections.Generic;
 
-namespace Read
+namespace Read.NationalSocietyFeatures
 {
     public class NationalSocieties : INationalSocieties
 	{
