+++ Source/Admin/Read/NationalSocietyFeatures/NationalSociety.cs
@@ -4,18 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 using System;
+using MongoDB.Bson.Serialization.Attributes;
 
-namespace Read
+namespace Read.NationalSocietyFeatures
 {
-    public class User
+    public class NationalSociety
     {
-        public Guid Id { get; set; }
-
-        public string Firstname { get; set; }
-
-        public string Lastname { get; set; }
-
+		public Guid Id { get; set; }
+		public string Name { get; set; }
         public string Country { get; set; }
-        public Guid NationalSocietyId { get; set; }
     }
 }
