+++ Source/Admin/Read/NationalSocietyFeatures/NationalSocietyCreatedProcessor.cs
@@ -1,4 +1,4 @@
-﻿/*---------------------------------------------------------------------------------------------
+/*---------------------------------------------------------------------------------------------
  *  Copyright (c) 2017 International Federation of Red Cross. All rights reserved.
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
@@ -8,7 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Read
+namespace Read.NationalSocietyFeatures
 {
 	public class NationalSocietyCreatedProcessor : Infrastructure.Events.IEventProcessor
 	{
