+++ Source/Admin/Read/ProjectFeatures/Projects.cs
@@ -3,61 +3,51 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-using MongoDB.Driver;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using MongoDB.Driver;
 
-namespace Read
+namespace Read.ProjectFeatures
 {
-    public class Users : IUsers
+    public class Projects : IProjects
     {
-        readonly IMongoCollection<User> _collection;
+        readonly IMongoCollection<Project> _collection;
 
         readonly IMongoDatabase _database;
 
-        public Users(IMongoDatabase database)
+        public Projects(IMongoDatabase database)
         {
             _database = database;
-            _collection = database.GetCollection<User>("User");
+            _collection = database.GetCollection<Project>("Project");
         }
 
-        public User GetById(Guid id)
+        public Project GetById(Guid id)
         {
             var project = _collection.Find(v => v.Id == id).SingleOrDefault();
-
             if (project == null)
             {
-                project = new User { Firstname = "Dummy implementation" };
+                project = new Project {Name = "Dummy implementation"};
                 _collection.InsertOne(project);
             }
-
             return project;
         }
 
-        public async Task<IEnumerable<User>> GetByNationalSocietyId(Guid id)
+        public void Save(Project project)
         {
-            var userList = await _collection.FindAsync(_ => _.NationalSocietyId == id);
-            return await userList.ToListAsync();
+            var filter = Builders<Project>.Filter.Eq(v => v.Id, project.Id);
+            _collection.ReplaceOne(filter, project);
         }
 
-        public void Save(User user)
-        {
-            var filter = Builders<User>.Filter.Eq(v => v.Id, user.Id);
-
-            _collection.ReplaceOne(filter, user);
-        }
-
-        public IEnumerable<User> GetAll()
+        public IEnumerable<Project> GetAll()
         {
             return _collection.Find(_ => true).ToList();
         }
 
-        public async Task<IEnumerable<User>> GetAllASync()
+        public async Task<IEnumerable<Project>> GetAllASync()
         {
-            var filter = Builders<User>.Filter.Empty;
+            var filter = Builders<Project>.Filter.Empty;
             var list = await _collection.FindAsync(filter);
-
             return await list.ToListAsync();
         }
     }
