+++ Source/Admin/Read/UserFeatures/User.cs
@@ -1,17 +1,21 @@
-﻿/*---------------------------------------------------------------------------------------------
+/*---------------------------------------------------------------------------------------------
  *  Copyright (c) 2017 International Federation of Red Cross. All rights reserved.
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
 using System;
-using MongoDB.Bson.Serialization.Attributes;
 
-namespace Read
+namespace Read.UserFeatures
 {
-    public class NationalSociety
+    public class User
     {
-		public Guid Id { get; set; }
-		public string Name { get; set; }
+        public Guid Id { get; set; }
+
+        public string Firstname { get; set; }
+
+        public string Lastname { get; set; }
+
         public string Country { get; set; }
+        public Guid NationalSocietyId { get; set; }
     }
 }
