+++ Source/Admin/Read/UserFeatures/UserEventProcessor.cs
@@ -7,7 +7,7 @@ using Events;
 using Events.External;
 using Infrastructure.Events;
 
-namespace Read
+namespace Read.UserFeatures
 {
     public class UserEventProcessor : IEventProcessor
     {
