+++ Source/Admin/Web/ProjectController.cs
@@ -10,6 +10,7 @@ using Infrastructure.Events;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Read;
+using Read.ProjectFeatures;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
