+++ Source/Admin/Web/TestDataGeneratorController.cs
@@ -10,6 +10,8 @@ using Read;
 using MongoDB.Driver;
 using Newtonsoft.Json;
 using Events.External;
+using Read.NationalSocietyFeatures;
+using Read.UserFeatures;
 
 namespace Web
 {
